refactor: remove logic making db instance everytime + move business logic(worker -> handler)
